Move Vaillant providers to public packages
