+++ providers/vaillant/providers.go
@@ -2,10 +2,10 @@ package vaillant
 
 import (
 	"github.com/d3vi1/helianthus-ebusreg/registry"
+	"github.com/d3vi1/helianthus-ebusreg/vaillant/dhw"
+	"github.com/d3vi1/helianthus-ebusreg/vaillant/heating"
+	"github.com/d3vi1/helianthus-ebusreg/vaillant/solar"
 	"github.com/d3vi1/helianthus-ebusreg/vaillant/system"
-	"github.com/d3vi1/helianthus-ebusreg/vendor/vaillant/dhw"
-	"github.com/d3vi1/helianthus-ebusreg/vendor/vaillant/heating"
-	"github.com/d3vi1/helianthus-ebusreg/vendor/vaillant/solar"
 )
 
 func System() registry.PlaneProvider {
