+++ vaillant/solar/schemas.go
