+++ vaillant/solar/schemas_test.go
