+++ vaillant/solar/solar.go
