+++ vaillant/solar/solar_test.go
